requirements.txt
streamlit==1.35.0
langchain-groq==0.1.2
langchain-core==0.2.1
sentence-transformers==2.7.0
chromadb==0.4.24
beautifulsoup4==4.12.3
python-dotenv==1.0.1
PyPDF2==3.0.1
requests==2.31.0
